+++ app/src/main/java/eu/kanade/tachiyomi/ui/manga/MangaScreenModel.kt
@@ -111,16 +111,16 @@ import tachiyomi.domain.library.service.LibraryPreferences
 import tachiyomi.domain.manga.interactor.DeleteByMergeId
 import tachiyomi.domain.manga.interactor.DeleteMangaById
 import tachiyomi.domain.manga.interactor.DeleteMergeById
-import tachiyomi.domain.external_watcher.interactor.RemoveFromExternalWatcher
+import tachiyomi.domain.watcher.interactor.RemoveFromExternalWatcher
 import tachiyomi.domain.manga.interactor.GetDuplicateLibraryManga
 import tachiyomi.domain.manga.interactor.GetFlatMetadataById
 import tachiyomi.domain.manga.interactor.GetManga
 import tachiyomi.domain.manga.interactor.GetMangaWithChapters
 import tachiyomi.domain.manga.interactor.GetMergedMangaById
 import tachiyomi.domain.manga.interactor.GetMergedReferencesById
-import tachiyomi.domain.external_watcher.interactor.GetExternalWatcher
+import tachiyomi.domain.watcher.interactor.GetExternalWatcher
 import tachiyomi.domain.manga.interactor.InsertMergedReference
-import tachiyomi.domain.external_watcher.interactor.AddToExternalWatcher
+import tachiyomi.domain.watcher.interactor.AddToExternalWatcher
 import tachiyomi.domain.manga.interactor.NetworkToLocalManga
 import tachiyomi.domain.manga.interactor.SetCustomMangaInfo
 import tachiyomi.domain.manga.interactor.SetMangaChapterFlags
@@ -130,7 +130,7 @@ import tachiyomi.domain.manga.model.Manga
 import tachiyomi.domain.manga.model.MangaUpdate
 import tachiyomi.domain.manga.model.MergeMangaSettingsUpdate
 import tachiyomi.domain.manga.model.MergedMangaReference
-import tachiyomi.domain.external_watcher.model.ExternalWatcherRequest
+import tachiyomi.domain.watcher.model.ExternalWatcherRequest
 import tachiyomi.domain.manga.model.applyFilter
 import tachiyomi.domain.manga.repository.MangaRepository
 import tachiyomi.domain.source.service.SourceManager
