+++ domain/src/main/java/tachiyomi/domain/watcher/ExternalWatcherHostProvider.kt
@@ -1,4 +1,4 @@
-package tachiyomi.domain.external_watcher
+package tachiyomi.domain.watcher
 
 const val EXTERNAL_WATCHER_HOST_DISABLED = ""
 const val EXTERNAL_WATCHER_HOST_ACHMAD = "https://api.achmad.dev"
